commit code

+++ seo-information-manager/src/main/java/com/yxy/dch/seo/information/web/ArticleController.java
@@ -7,17 +7,22 @@ import com.yxy.dch.seo.information.config.pros.MinioProperties;
 import com.yxy.dch.seo.information.exception.BizException;
 import com.yxy.dch.seo.information.exception.CodeMsg;
 import com.yxy.dch.seo.information.service.IArticleService;
+import com.yxy.dch.seo.information.util.JacksonUtil;
+import com.yxy.dch.seo.information.util.Toolkit;
 import com.yxy.dch.seo.information.vo.ArticleVO;
 import com.yxy.dch.seo.information.vo.ColumnVO;
+import com.yxy.dch.seo.information.vo.ImgVO;
 import com.yxy.dch.seo.information.vo.Page;
 import io.minio.MinioClient;
 import io.swagger.annotations.Api;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.InputStream;
 import java.util.List;
 
 /**
@@ -42,11 +47,18 @@ public class ArticleController extends BaseController {
      */
     @PostMapping("/create")
     public ArticleVO create(ArticleVO param) {
+        if (StringUtils.isBlank(param.getName()) || param.getColumnId() == null || param.getRecommend() == null || StringUtils.isBlank(param.getContent())) {
+            logger.error("新增文章参数错误({}):param={}",CodeMsg.param_note_blank.getMsg(),JacksonUtil.toJson(param));
+            throw new BizException(CodeMsg.param_note_blank);
+        }
+        logger.info("新增文章:param={}", JacksonUtil.toJson(param));
         // 操作用户ID
         Long opeUid = UserReqContextUtil.getRequestUserId();
         param.setCreateUid(opeUid);
         param.setUpdateUid(opeUid);
-        return articleService.create(param);
+        ArticleVO articleVO = articleService.create(param);
+        logger.info("新增文章成功,result={}", JacksonUtil.toJson(articleVO));
+        return articleVO;
     }
 
     /**
@@ -57,7 +69,14 @@ public class ArticleController extends BaseController {
      */
     @PostMapping("/remove")
     public Boolean remove(ArticleVO param) {
-        return articleService.remove(param);
+        if (param.getId() == null){
+            logger.error("删除文章参数错误({}):param={}",CodeMsg.id_param_blank.getMsg(),JacksonUtil.toJson(param));
+            throw new BizException(CodeMsg.id_param_blank);
+        }
+        logger.info("删除文章:param={}", JacksonUtil.toJson(param));
+        Boolean ret = articleService.remove(param);
+        logger.info("删除文章成功:result={}", ret);
+        return ret;
     }
 
     /**
@@ -68,7 +87,14 @@ public class ArticleController extends BaseController {
      */
     @PostMapping("/modify")
     public ArticleVO modify(ArticleVO param) {
-        return articleService.modify(param);
+        if (param.getId() == null){
+            logger.error("修改文章参数错误({}):param={}",CodeMsg.id_param_blank.getMsg(),JacksonUtil.toJson(param));
+            throw new BizException(CodeMsg.id_param_blank);
+        }
+        logger.info("修改文章:param={}", JacksonUtil.toJson(param));
+        ArticleVO articleVO = articleService.modify(param);
+        logger.info("修改文章成功:result={}", JacksonUtil.toJson(articleVO));
+        return articleVO;
     }
 
     /**
@@ -79,7 +105,14 @@ public class ArticleController extends BaseController {
      */
     @PostMapping("/view")
     public ArticleVO view(ArticleVO param) {
-        return articleService.view(param);
+        if (param.getId() == null){
+            logger.error("查看文章参数错误({}):param={}",CodeMsg.id_param_blank.getMsg(),JacksonUtil.toJson(param));
+            throw new BizException(CodeMsg.id_param_blank);
+        }
+        logger.info("查看文章:param={}", JacksonUtil.toJson(param));
+        ArticleVO articleVO = articleService.view(param);
+        logger.info("查看文章成功:result={}", JacksonUtil.toJson(articleVO));
+        return articleVO;
     }
 
     /**
@@ -91,8 +124,14 @@ public class ArticleController extends BaseController {
      */
     @PostMapping("/listByPage")
     public PageInfo<ArticleVO> listByPage(ArticleVO param, Page page) {
+        if (page == null || page.getPageSize() == null || page.getPageNum() == null){
+            logger.error("分页查询文章列表参数错误({}):param={},page={}",CodeMsg.user_batch_query_num_out.getMsg(),JacksonUtil.toJson(param),JacksonUtil.toJson(page));
+            throw new BizException(CodeMsg.user_batch_query_num_out);
+        }
+        logger.info("分页查询文章列表:param={},page={}", JacksonUtil.toJson(param), page);
         PageHelper.startPage(page.getPageNum(), page.getPageSize(), true);
         List<ArticleVO> list = articleService.listBy(param);
+        logger.info("分页查询文章列表成功:result={}", JacksonUtil.toJson(list));
         return new PageInfo<>(list);
     }
 
@@ -104,7 +143,14 @@ public class ArticleController extends BaseController {
      */
     @PostMapping("/up")
     public ArticleVO up(ArticleVO param) {
-        return articleService.up(param);
+        if (param.getId() == null){
+            logger.error("上架文章参数错误({}):param={}",CodeMsg.id_param_blank.getMsg(),JacksonUtil.toJson(param));
+            throw new BizException(CodeMsg.id_param_blank);
+        }
+        logger.info("上架文章:param={}", JacksonUtil.toJson(param));
+        ArticleVO articleVO = articleService.up(param);
+        logger.info("上架文章成功,result={}", JacksonUtil.toJson(articleVO));
+        return articleVO;
     }
 
     /**
@@ -115,7 +161,14 @@ public class ArticleController extends BaseController {
      */
     @PostMapping("/down")
     public ArticleVO down(ArticleVO param) {
-        return articleService.down(param);
+        if (param.getId() == null){
+            logger.error("下架文章参数错误({}):param={}",CodeMsg.id_param_blank.getMsg(),JacksonUtil.toJson(param));
+            throw new BizException(CodeMsg.id_param_blank);
+        }
+        logger.info("下架文章:param={}", JacksonUtil.toJson(param));
+        ArticleVO articleVO = articleService.down(param);
+        logger.info("下架文章成功,result={}", JacksonUtil.toJson(articleVO));
+        return articleVO;
     }
 
 
@@ -128,23 +181,82 @@ public class ArticleController extends BaseController {
      */
     @PostMapping("/uploadContent")
     public String uploadContent(ColumnVO param, MultipartFile file) {
+        if (param.getId() == null){
+            logger.error("上传文章内容参数错误({}):param={},file={}",CodeMsg.id_param_blank.getMsg(),JacksonUtil.toJson(param),JacksonUtil.toJson(file));
+            throw new BizException(CodeMsg.id_param_blank);
+        }
+        logger.info("上传文章内容(articleId={},contentType={},filename={},size={})", param.getId(), file.getContentType(),file.getOriginalFilename(),file.getSize());
         String pictureUrl;
         try {
             MinioClient minioClient = new MinioClient(minioProperties.getEndpoint(), minioProperties.getAccessKey(), minioProperties.getSecretKey());
             String contentType = file.getContentType();
             String originalFilename = file.getOriginalFilename();
             String bucketName = minioProperties.getBucketName();
-            String articleContentDir = minioProperties.getArticleContentDir();
-            String objectName = articleContentDir + "/" + param.getId() + "/" + originalFilename;
+            String articleContentBaseDir = minioProperties.getArticleContentBaseDir();
+//            String objectName = articleContentBaseDir + "/" + param.getId() + "/" + originalFilename;
+            String objectName = genObjectName(articleContentBaseDir,originalFilename,param.getId());
             boolean isExist = minioClient.bucketExists(bucketName);
             if (!isExist) {
                 minioClient.makeBucket(bucketName);
             }
-            minioClient.putObject(bucketName, objectName, file.getInputStream(), contentType);
-            pictureUrl = minioClient.getObjectUrl(bucketName, objectName);
+            try(InputStream inputStream = file.getInputStream()){
+                minioClient.putObject(bucketName, objectName, inputStream, contentType);
+                pictureUrl = minioClient.getObjectUrl(bucketName, objectName);
+            }catch (Exception e){
+                throw new BizException(CodeMsg.system_error);
+            }
         } catch (Exception e) {
             throw new BizException(CodeMsg.system_error);
         }
+        logger.info("上传文章内容成功({})", pictureUrl);
         return pictureUrl;
+    }
+
+    /**
+     * 上传文章内容图片
+     * @param file 文章内容图片
+     * @return 内容图片
+     */
+    @PostMapping("/uploadImg")
+    public ImgVO uploadImg(MultipartFile file) {
+        if (file == null || file.isEmpty()){
+            logger.error("上传文章内容图片参数错误,文件为空");
+            throw new BizException(CodeMsg.param_note_blank);
+        }
+        logger.info("上传文章内容图片(contentType={},filename={},size={}", file.getContentType(),file.getOriginalFilename(),file.getSize());
+        String pictureUrl;
+        String contentType = file.getContentType();
+        String originalFilename = file.getOriginalFilename();
+        String bucketName = minioProperties.getBucketName();
+        String articleContentImgBaseDir = minioProperties.getArticleContentImgBaseDir();
+//        String objectName = articleContentImgBaseDir + "/" + originalFilename;
+        String objectName = genObjectName(articleContentImgBaseDir,originalFilename,null);
+        try {
+            MinioClient minioClient = new MinioClient(minioProperties.getEndpoint(), minioProperties.getAccessKey(), minioProperties.getSecretKey());
+            boolean isExist = minioClient.bucketExists(bucketName);
+            if (!isExist) {
+                minioClient.makeBucket(bucketName);
+            }
+            try(InputStream inputStream = file.getInputStream()){
+                minioClient.putObject(bucketName, objectName, inputStream, contentType);
+                pictureUrl = minioClient.getObjectUrl(bucketName, objectName);
+            }catch (Exception e){
+                throw new BizException(CodeMsg.system_error);
+            }
+        } catch (Exception e) {
+            throw new BizException(CodeMsg.file_system_error);
+        }
+        ImgVO imgVO = new ImgVO();
+        imgVO.setSrc(pictureUrl);
+        imgVO.setTitle(originalFilename);
+        logger.info("上传文章内容图片成功({})", JacksonUtil.toJson(imgVO));
+        return imgVO;
+    }
+
+    private String genObjectName(String baseDir, String originalFilename, Long businessId) {
+        if (businessId != null){
+            return baseDir + "/" + businessId + "/" + Toolkit.randomUUID();
+        }
+        return baseDir + "/" + originalFilename + "/" + Toolkit.randomUUID();
     }
 }
